Rewrite the lwt stub using an internal thread
Using lwt.preemptive requires downstream users to compile with `-thread`
which seems egregious. Unfortunately the lwt job interface headers are
not guaranteed to be installed anywhere in particular (eg somewhere
under .opam, not on the include path).

Instead we fork a background pthread in stub, and return a pipe fd to
OCaml. We marshal the results over the pipe and receive them in OCaml
code.

Signed-off-by: David Scott <dave.scott@unikernel.com>